Exercises corrected and lesson 8 completed.

+++ Exercise07/exercise07.py
@@ -1,6 +1,6 @@
 studentName = input('What is the name of the student?')
-firstNote = int(input('What is the first note of {}?' .format(studentName)))
-secondNote = int(input('What is the second note?'))
+firstNote = float(input('What is the first note of {}?' .format(studentName)))
+secondNote = float(input('What is the second note?'))
 average = (firstNote + secondNote) / 2
 
 print('The average of the notes of the student {} is: {} ' .format(studentName, average))
